Predictor/requirements.txt
requests
psycopg2-binary
pandas
numpy
scikit-learn
matplotlib
seaborn